requirements.txt
torch>=2.0.0
torchvision>=0.15.0

transformers>=4.35.0
accelerate>=0.20.0
peft>=0.6.0

Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0

PyYAML>=6.0
tqdm>=4.65.0

hydra-core>=1.3.0
omegaconf>=2.3.0

loguru>=0.7.0